Update README.md
​
24
If the resultant hashes match, the bid is valid. The contract will track the highest bid as the reveal()
25
function is called by each participant during the reveal phase. Once all reveals are complete, the highest
26
bidder can be determined on chain and an order can be signed and filled to complete the transaction.
27
​
28
## Why Blockchain?
29
​
30
A blockchain is a perfect medium for recording commitments to some value that may be revealed at a later 
31
time or state. So if designed correctly, it should be very difficult for auction participants to collude
32
with each other or otherwise sabotage the auction. At the same time, anyone in the world could adjudicate
33
on the fairness of the auction if provided with an honest record of those commitments and revealed bids.
34
​
35
Even more importantly, this kind of auction is essentially a game of incomplete information amongst bidders 
36
with interesting game theoretic properties like Nash equilibria. They may be useful in price discovery for 
37
difficult to price goods or services. If decentralization will usher in new types of previously unimagined 
38
digital assets, it seems natural that effective and fair price discovery for both buyers and sellers would 
39
be a worthy area of research.
40
​
41
## State Channel